+++ lib/pages/settings/dashboards/dashboard_item_card.dart
@@ -189,7 +189,7 @@ class ItemCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Card(
-      color: styleConfig().secondaryTextColor,
+      color: styleConfig().secondaryTextColor.withOpacity(0.3),
       elevation: 8,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(8),