+++ lib/widgets/charts/habits/dashboard_habits_chart.dart
@@ -166,7 +166,6 @@ class HabitChartInfoWidget extends StatelessWidget {
                 IconButton(
                   padding: const EdgeInsets.symmetric(horizontal: 10),
                   onPressed: onTapAdd,
-                  hoverColor: Colors.transparent,
                   icon: SvgPicture.asset(styleConfig().addIcon),
                 ),
               ],