+++ lib/widgets/journal/entry_details/share_button_widget.dart
@@ -56,7 +56,6 @@ class ShareButtonWidget extends StatelessWidget {
           width: 40,
           child: IconButton(
             icon: const Icon(MdiIcons.shareOutline),
-            hoverColor: Colors.transparent,
             splashColor: Colors.transparent,
             iconSize: 24,
             tooltip: tooltip,