+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Navigate back icon on task page
 - Unordered list color in editor
 - Editor menu background color
+- Chip style
 
 ### Changed:
 - Improved task input fields layout
 - Improved chip layout
 - Tweak spacing in journal filters
+- Improved styling in settings
 
 ## [0.8.264] - 2023-02-11
 ### Fixed: